+++ src/main/java/Game/Map/Square.java
@@ -1,6 +1,12 @@
 package Game.Map;
 
 public interface Square {
-    public int getWalkSpeed();
+    /*
+     * How much of a move difference does it give?, rounds down
+     */
+    public double getWalkSpeed();
+    /*
+     *  Percentage you take 1 health of damage
+     */
     public double getDanger();
 }
